Refine the reporter API
All of the assertion methods have been consolidated into a single assert
method that takes a guard and parallel message and object arrays and
allows the reporter to interleave them as it sees fit.

And consolidate the Node.js and Phantom to Node.js proxy reporters into
one module with a few conditional differences. This is facilitated by
using the Browser `util` module port.

Adds support for toBeInRange and toBeBetween