Pipe color through the runner
Unless the Node.js console is not a TTY, both for the Node.js and
PhantomJS test runners.

Show passed tests with the `-p` or `--passed` options.

Fix a bug for an empty list of tests.

Allow descriptions to be skipped by omitting the callback.

Pass options through to `Reporter` constructor in `runAndReport`.